+++ ci/gpu/build.sh
@@ -43,7 +43,10 @@ nvidia-smi
 logger "Activate conda env..."
 source activate gdf
 conda install "rmm=$MINOR_VERSION.*" "nvstrings=$MINOR_VERSION.*" "cudatoolkit=$CUDA_REL" \
-              "dask>=2.0" "distributed>=2.0" "numpy>=1.16" "dlpack" "fastavro>=0.22.0"
+              "dask>=2.1.0" "distributed>=2.1.0" "numpy>=1.16" "double-conversion" \
+              "rapidjson" "flatbuffers" "boost-cpp" "fsspec>=0.3.3" "dlpack" \
+              "feather-format" "cupy>=6.0.0" "arrow-cpp=0.14.1" "pyarrow=0.14.1" \
+              "fastavro>=0.22.0" "pandas>=0.24.2,<0.25"
 
 # Install the master version of dask and distributed
 logger "pip install git+https://github.com/dask/distributed.git --upgrade --no-deps" 
@@ -77,14 +80,6 @@ else
     logger "GoogleTest for libcudf..."
     cd $WORKSPACE/cpp/build
     GTEST_OUTPUT="xml:${WORKSPACE}/test-results/" make -j${PARALLEL_LEVEL} test
-
-    # Install the master version of distributed for serialization testing
-    logger "pip install git+https://github.com/dask/distributed.git"
-    pip install "git+https://github.com/dask/distributed.git"
-
-    # Temporarily install feather and cupy for testing
-    logger "conda install feather-format"
-    conda install "feather-format" "cupy>=6.0.0"
 
     # set environment variable for numpy 1.16
     # will be enabled for later versions by default