+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - PR #2345 Restore removal of old groupby implementation
 - PR #2329 using libcudf cudf::copy for column deep copy
 - PR #2344 Add docs on how code formatting works for contributors
+- PR #2353 Bump Arrow and Dask versions
 - PR #2377 Replace `standard_python_slice` with just `slice.indices()`
 - PR #2392 Remove dlpack submodule; make cuDF's Cython API externally accessible
 
@@ -98,6 +99,7 @@
 - PR #2361 Fixed issue with `codes` of CategoricalIndex
 - PR #2357 Fixed inconsistent type of index created with from_pandas vs direct construction
 - PR #2389 Fixed Rolling __getattr__ and __getitem__ for offset based windows
+- PR #2401 Fix to a scalar datetime(of type Days) issue
 
 
 # cuDF 0.8.0 (27 June 2019)